Move QueryStatusReportListeners to new query-runner module
