+++ java/clients/pom.xml
@@ -83,7 +83,7 @@
         <!-- Sleeper dependencies -->
         <dependency>
             <groupId>sleeper</groupId>
-            <artifactId>query-core</artifactId>
+            <artifactId>query-runner</artifactId>
             <version>${project.parent.version}</version>
         </dependency>
         <dependency>
