+++ java/clients/src/main/java/sleeper/clients/QueryLambdaClient.java
@@ -29,8 +29,8 @@ import sleeper.configuration.properties.table.TablePropertiesProvider;
 import sleeper.core.statestore.StateStoreException;
 import sleeper.query.model.Query;
 import sleeper.query.model.QuerySerDe;
-import sleeper.query.model.output.ResultsOutputConstants;
-import sleeper.query.model.output.SQSResultsOutput;
+import sleeper.query.output.ResultsOutputConstants;
+import sleeper.query.output.SQSResultsOutput;
 import sleeper.query.tracker.DynamoDBQueryTracker;
 import sleeper.query.tracker.QueryState;
 import sleeper.query.tracker.TrackedQuery;
