+++ java/query/query-core/src/main/java/sleeper/query/output/ResultsOutputLocation.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.query.model.output;
+package sleeper.query.output;
 
 public class ResultsOutputLocation {
     private final String type;
