+++ java/query/query-core/src/main/java/sleeper/query/tracker/DynamoDBQueryTracker.java
@@ -31,7 +31,7 @@ import org.slf4j.LoggerFactory;
 import sleeper.configuration.properties.instance.InstanceProperties;
 import sleeper.query.model.LeafPartitionQuery;
 import sleeper.query.model.Query;
-import sleeper.query.model.output.ResultsOutputInfo;
+import sleeper.query.output.ResultsOutputInfo;
 import sleeper.query.tracker.exception.QueryTrackerException;
 
 import java.util.List;
