+++ java/query/query-core/src/main/java/sleeper/query/tracker/DynamoDBQueryTrackerEntry.java
@@ -22,7 +22,7 @@ import com.amazonaws.services.dynamodbv2.model.AttributeValueUpdate;
 
 import sleeper.query.model.LeafPartitionQuery;
 import sleeper.query.model.Query;
-import sleeper.query.model.output.ResultsOutputInfo;
+import sleeper.query.output.ResultsOutputInfo;
 
 import java.util.HashMap;
 import java.util.Map;
