+++ java/query/query-core/src/test/java/sleeper/query/model/output/WebSocketResultsOutputIT.java
@@ -24,6 +24,8 @@ import sleeper.core.iterator.WrappedIterator;
 import sleeper.core.record.Record;
 import sleeper.query.model.Query;
 import sleeper.query.model.QueryOrLeafPartitionQuery;
+import sleeper.query.output.ResultsOutputInfo;
+import sleeper.query.output.WebSocketResultsOutput;
 
 import java.util.ArrayList;
 import java.util.Collections;
