+++ java/query/query-lambda/src/main/java/sleeper/query/lambda/QueryMessageHandler.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022-2023 Crown Copyright
+ * Copyright 2022-2024 Crown Copyright
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,8 +23,8 @@ import sleeper.configuration.properties.table.TablePropertiesProvider;
 import sleeper.query.model.QueryOrLeafPartitionQuery;
 import sleeper.query.model.QuerySerDe;
 import sleeper.query.model.QueryValidationException;
+import sleeper.query.runner.tracker.QueryStatusReportListeners;
 import sleeper.query.tracker.QueryStatusReportListener;
-import sleeper.query.tracker.QueryStatusReportListeners;
 
 import java.util.Optional;
 import java.util.UUID;
