+++ java/query/query-lambda/src/test/java/sleeper/query/lambda/SqsQueryProcessorLambdaIT.java
@@ -72,10 +72,10 @@ import sleeper.io.parquet.record.ParquetRecordReader;
 import sleeper.query.model.Query;
 import sleeper.query.model.QueryProcessingConfig;
 import sleeper.query.model.QuerySerDe;
-import sleeper.query.model.output.ResultsOutputConstants;
-import sleeper.query.model.output.S3ResultsOutput;
-import sleeper.query.model.output.SQSResultsOutput;
-import sleeper.query.model.output.WebSocketResultsOutput;
+import sleeper.query.output.ResultsOutputConstants;
+import sleeper.query.output.S3ResultsOutput;
+import sleeper.query.output.SQSResultsOutput;
+import sleeper.query.output.WebSocketResultsOutput;
 import sleeper.query.tracker.DynamoDBQueryTracker;
 import sleeper.query.tracker.DynamoDBQueryTrackerCreator;
 import sleeper.query.tracker.QueryStatusReportListener;
