+++ java/query/query-runner/src/test/java/sleeper/query/runner/tracker/WebSocketQueryStatusReportDestinationIT.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022-2023 Crown Copyright
+ * Copyright 2022-2024 Crown Copyright
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.query.tracker;
+package sleeper.query.runner.tracker;
 
 import com.amazonaws.auth.BasicAWSCredentials;
 import com.github.tomakehurst.wiremock.junit5.WireMockRuntimeInfo;
@@ -32,8 +32,9 @@ import sleeper.core.schema.type.LongType;
 import sleeper.core.schema.type.StringType;
 import sleeper.query.model.LeafPartitionQuery;
 import sleeper.query.model.Query;
-import sleeper.query.model.output.ResultsOutputInfo;
-import sleeper.query.model.output.ResultsOutputLocation;
+import sleeper.query.output.ResultsOutputInfo;
+import sleeper.query.output.ResultsOutputLocation;
+import sleeper.query.tracker.WebSocketQueryStatusReportDestination;
 
 import java.io.IOException;
 import java.util.List;
